pyproject.toml
[project]
name = "tabu-self-driving-car"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "maks-tabu", email = "maxtabolin@gmail.com"},
]
dependencies = [
    "polars==1.12.0", 
    "catboost==1.2.7", 
    "pytorch-tsmixer==0.2.0", 
    "scikit-learn==1.5.2", 
    "tqdm==4.66.6", 
    "typer==0.12.5", 
    "black==24.10.0", 
    "pyarrow==18.0.0"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
